+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 - FIX : Identities enumeration fix
 - FIX : typos
-- Exporting only non-expired Certificates
 - FIX : Certificate label is force converted to UTF8 before using it, to not to break in case there's a non UTF8 character
+- Don't fail, just print warning if multiple Identities found for a single search (e.g. in case you have previous, revoked versions of the Certificate in your Keychain)
+- Certificate filtering: export only non-expired, valid by date Certificates
 
 
 ## 0.9.4